+++ source/editor/gui/tab/scene-editor/transform/gizmo/TransformGizmoScale.js
@@ -1,4 +1,3 @@
-import {Shape} from "cannon-es";
 import {Geometry, Mesh, BoxGeometry, BufferGeometry, Float32BufferAttribute, Line, BoxBufferGeometry, CylinderBufferGeometry, Matrix4} from "three";
 import {ChangeAction} from "../../../../../history/action/ChangeAction.js";
 import {ActionBundle} from "../../../../../history/action/ActionBundle.js";