+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### V0.108
 
 - Removed animation related functionality from editor
+- Removed cannon-es library support completely from editor
 
 ### V0.107